Tests/LibWeb/Text/expected/github-split-diff-fixed-table-width.txt

table width: 600
cell widths: 150, 150, 150, 150
cell width sum: 600
